Fixed detection of correct process
With using grep (in Linux) or find (in Windows) without boundaries it could happen that the wrong process was found (for example if the PID was part of an open port of another application). By using boundaries (space before and line-end/slash after the PID) this mismatch can no longer happen.

By coincidence using findstr instead of find in Windows also added automatic cygwin support because there is no longer a collision for the find command.