Fix `lookup()` helper
Accessing properties of Phantom's `page` object doesn't return a ref to the
original object (there must be some kind of getter). As a result,
deep-setting subproperties has no effect.

This fix changes the semantics of the helper: when setting a value, we don't
return value but a reference to the object for overriding.

Bonus: allow deep-setting a prop when its parents dont exist by creating empty
object on the fly.